CHORE: Initial commit for a working response
Response works successfully on passing the city name to the url endpoint "localhost:8080/weather?city={name}'
To do:
* Persisting the response with a proper pojo

+++ weather/src/main/java/weather/Weather.java
@@ -49,7 +49,7 @@ public class Weather {
         this.cod = cod;
     }
 
-    public String toString(String city, Double min, Double max, boolean carryUmbrella, boolean useSunScreen) {
+    public String toString(String city, Long min, Long max, boolean carryUmbrella, boolean useSunScreen) {
         return "Weather forecast for the city "+city +
                 " for the next three days: \n\nMaximum temperature(In kelvin) = '" + max + '\'' +
                 "\nMinimum temperature(In kelvin) = '" + min +'\''+
